♻️  core: Split proxy mod
